+++ src/main/java/codesquad/fineants/domain/oauth/decoder/KakaoIDTokenDecoder.java
@@ -1,4 +1,4 @@
-package codesquad.fineants.spring.api.member.decoder;
+package codesquad.fineants.domain.oauth.decoder;
 
 import codesquad.fineants.domain.oauth.client.DecodedIdTokenPayload;
 import codesquad.fineants.domain.oauth.client.kakao.KakaoDecodedIdTokenPayload;
