+++ src/main/java/codesquad/fineants/domain/oauth/properties/OauthProperties.java
@@ -10,8 +10,8 @@ import codesquad.fineants.domain.oauth.client.OauthClient;
 import codesquad.fineants.domain.oauth.client.google.GoogleOauthClient;
 import codesquad.fineants.domain.oauth.client.kakao.KakaoOauthClient;
 import codesquad.fineants.domain.oauth.client.naver.NaverOauthClient;
-import codesquad.fineants.spring.api.member.decoder.GoogleIDTokenDecoder;
-import codesquad.fineants.spring.api.member.decoder.KakaoIDTokenDecoder;
+import codesquad.fineants.domain.oauth.decoder.GoogleIDTokenDecoder;
+import codesquad.fineants.domain.oauth.decoder.KakaoIDTokenDecoder;
 import codesquad.fineants.spring.api.member.service.WebClientWrapper;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
